modify infinity scroll product
