+++ src/main/java/kr/co/rice/controller/ProductListController.java
@@ -1,5 +1,9 @@
 package kr.co.rice.controller;
 
+import java.util.ArrayList;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
@@ -9,8 +13,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import kr.co.rice.service.ProductListService;
+import kr.co.rice.vo.ProductVo;
+
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 public class ProductListController {
@@ -22,5 +29,12 @@ public class ProductListController {
 	@RequestMapping("/products/{category}/{subcategory}")
 	public String category_list(@PathVariable("category")int category_id, @PathVariable("subcategory") int subcategory_id, Model model) {
 		return service.category_list(category_id, subcategory_id, model);
+	}
+	
+	@ResponseBody
+	@RequestMapping(value="/products/getItems",method=RequestMethod.GET,produces = "application/json; charset=utf8")
+	public ArrayList<ProductVo> getItems(HttpServletRequest request ,Model model)
+	{
+		return service.getItems(request,model);
 	}
 }
