+++ src/main/java/kr/co/rice/mapper/ProductListMapper.java
@@ -6,5 +6,5 @@ import kr.co.rice.vo.ProductVo;
 
 public interface ProductListMapper {
 	public ArrayList<ProductVo> category_list(int category_id, int subcategory_id);
-	public ArrayList<ProductVo> getItems(int page);
+	public ArrayList<ProductVo> getItems(int category_id,int subcategory_id,int start, int end);
 }
