+++ src/main/java/kr/co/rice/service/SearchServiceImpl.java
@@ -34,16 +34,16 @@ public class SearchServiceImpl implements SearchService{
 		return "/products/list";
 	}
 	
-	@Override
-	public ArrayList<ProductVo> getItems(HttpServletRequest request,Model model) {
-		String search = request.getParameter("search");
-		int start = Integer.parseInt(request.getParameter("start"));
-		int end = Integer.parseInt(request.getParameter("end"));
-		
-		ArrayList<ProductVo>list = mapper.getItems(search,start,end);
-		
-		
-		return list;
-	}
+//	@Override
+//	public ArrayList<ProductVo> getItems(HttpServletRequest request,Model model) {
+//		String search = request.getParameter("search");
+//		int start = Integer.parseInt(request.getParameter("start"));
+//		int end = Integer.parseInt(request.getParameter("end"));
+//		
+//		ArrayList<ProductVo>list = mapper.getItems(search,start,end);
+//		
+//		
+//		return list;
+//	}
 	
 }
