+++ src/main/resources/kr/co/rice/mapper/ProductListMapper.xml
@@ -9,6 +9,17 @@
 		inner join categories as c on c.id = sc.category_id
 		inner join products_images as pi on pi.product_id = p.id
 		where sc.category_id = #{param1} and p.subcategory_id = #{param2}
+		limit 3
+	</select>
+	
+	<select id="getItems" resultType="kr.co.rice.vo.ProductVo">
+		select sl.id as spicylevel_id, sl.level as spicylevel, sc.name as subcategory_name,
+		p.id, p.title, p.name, p.size, p.price, pi.image_url from products as p
+		inner join products_images as pi on pi.product_id = p.id
+		inner join subcategories as sc on p.subcategory_id = sc.id
+		inner join spicylevels as sl on p.spicylevel_id = sl.id
+		where sc.category_id = #{param1} and p.subcategory_id = #{param2}
+		order by p.id desc limit #{param3},#{param4}
 	</select>
 	
 </mapper>
